changed syling

+++ trv-venue-hotel-client/src/app/administration/Administration.jsx
@@ -4,6 +4,13 @@ import axios from 'axios';
 import Hotel from './hotel/Hotel';
 import Link from '../commonComponents/Link';
 import Container from '../commonComponents/Container';
+import ButtonsContainer from '../commonComponents/ButtonsContainer';
+
+const Content = styled.div`
+margin: 1em auto;
+max-width: 980px;
+position: relative;
+`;
 
 export default class Administration extends Component {
 	state = {
@@ -27,7 +34,11 @@ export default class Administration extends Component {
 			});
 	}
 	handleRemove = (id) => {
-		alert('remove ' + id);
+		const name = this.state.hotels.find(hotel => hotel.id === id).name;
+		const confirmation = window.confirm("are you shure that you want to delete \""+name+"\"");
+		if(!confirmation){
+			return undefined;
+		}
 		axios.delete('/hotels/'+id)
 		.then((response) => {
 			this.handleFetch();
@@ -45,7 +56,13 @@ export default class Administration extends Component {
 		}
 		return (
 			<Container>
-				<Link to={this.props.match.path + '/new-hotel'}>add a new hotel</Link>
+				<Content>
+					<ButtonsContainer>
+					<Link to={this.props.match.path + '/new-hotel'}>add a new hotel</Link>
+				
+					</ButtonsContainer>
+				
+				</Content>
 				{
 					hotels.map(hotel => (
 								<Hotel key={hotel.id} hotel={hotel} onRemove={this.handleRemove}/>
